+++ backend-comparison/README.md
@@ -1,7 +1,7 @@
 # Burn Benchmark
 
 This crate is used with `cargo bench --features <backend>`
-to compare backend computations, from tensor operations to complex models.
+to compare backend computation times, from tensor operations to complex models.
 
 Note: in order to compare different backend-specific tensor operation
 implementations (for autotuning purposes, for instance), this should be done
