--- burn-wgpu/src/benchmark/mod.rs
@@ -1,3 +1,0 @@
-mod matmul;
-
-pub use matmul::*;
